add documentation, move dataloader to own folder
