+++ grants_tagger_light/training/train.py
@@ -1,73 +1,14 @@
-from transformers import AutoTokenizer, Trainer, TrainingArguments, EvalPrediction
-from datasets import Dataset
+from transformers import (
+    AutoTokenizer,
+    Trainer,
+    TrainingArguments,
+    EvalPrediction,
+)
 from grants_tagger_light.models.bert_mesh import BertMesh
-import json
+from grants_tagger_light.training.dataloaders import load_grants_sample
 import typer
 import numpy as np
 from sklearn.metrics import classification_report
-
-
-def load_data(
-    data_path: str,
-    tokenizer: AutoTokenizer,
-    label2id: dict,
-    test_size: float = 0.1,
-    num_proc: int = 8,
-):
-    def _datagen(data_path: str):
-        """
-        Loads the data from the given path. The data should be in jsonl format,
-        with each line containing a text and tags field.
-        The tags field should be a list of strings.
-        """
-        with open(data_path, "r") as f:
-            for line in f:
-                sample = json.loads(line)
-                yield sample
-
-    def _tokenize(batch):
-        return tokenizer(
-            batch["text"], padding="max_length", truncation=True, max_length=512
-        )
-
-    def _label_encode(batch):
-        batch["labels"] = [
-            [label2id[tag] for tag in tags if tag in label2id] for tags in batch["tags"]
-        ]
-        return batch
-
-    def _one_hot(batch):
-        batch["labels"] = [
-            [1 if i in labels else 0 for i in range(len(label2id))]
-            for labels in batch["labels"]
-        ]
-        return batch
-
-    dset = Dataset.from_generator(_datagen, gen_kwargs={"data_path": data_path})
-    dset = dset.map(
-        _tokenize, batched=True, batch_size=32, num_proc=num_proc, desc="Tokenizing"
-    )
-
-    dset = dset.map(
-        _label_encode,
-        batched=True,
-        batch_size=32,
-        num_proc=num_proc,
-        desc="Encoding labels",
-    )
-
-    dset = dset.map(
-        _one_hot,
-        batched=True,
-        batch_size=32,
-        num_proc=num_proc,
-        desc="One-hot labels",
-    )
-
-    # Split into train and test
-    dset = dset.train_test_split(test_size=test_size)
-
-    return dset["train"], dset["test"]
 
 
 def train_bertmesh(model_key: str, data_path: str, **user_args):
@@ -76,7 +17,7 @@ def train_bertmesh(model_key: str, data_path: str, **user_args):
 
     label2id = {v: k for k, v in model.id2label.items()}
 
-    train_dset, val_dset = load_data(data_path, tokenizer, label2id=label2id)
+    train_dset, val_dset = load_grants_sample(data_path, tokenizer, label2id=label2id)
 
     training_args = {
         "output_dir": "model_output",
@@ -148,7 +89,7 @@ def train_bertmesh_cli(
     ),
     model_save_path: str = typer.Argument(..., help="Path to save model to"),
 ):
-    train_bertmesh(model_key, data_path, model_save_path)
+    train_bertmesh(model_key, data_path)
 
 
 if __name__ == "__main__":
